requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
google-auth==2.35.0
passlib==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.9.0
httpx==0.27.2
